Collapse service and resouce models directly into the client based on PR review.  Cleanup tests.
